requirements.txt
flask-cors
flask
langchain-openai
langchain
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.76
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.30
mysql==0.0.3
mysql-connector-python==9.4.0
mysqlclient==2.2.7
numpy==2.3.3
openai==1.109.0
openpyxl==3.1.5
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
xxhash==3.5.0
zstandard==0.25.0
